// src/models/android.settings.key.ts
export enum AndroidSettingsKey {
    DEVELOPER_OPTIONS = "development_settings_enabled",
    USB_DEBUGGING = "adb_enabled",
    USB_DEBUGGING_V2 = "adb_enabled_toggleable",
    STRICT_MODE = "package_verifier_enable",
    FORCE_RTL = "debug.force_rtl",
    BACKGROUND_LIMIT_PROCESS = "background_process_limit",
    GPU_RENDERING = "debug.hwui.renderer",
    SHOW_TOUCHES = "show_touches",
    HARDWARE_ACCELERATION = "debug.hwui.render_dirty_regions",
    LAYOUT_BOUNDS = "debug.layout",
    GPU_OVERDRAW = "debug.hwui.overdraw",
    OPEN_GL_TRACE = "debug.hwui.profile",
    TOGGLE_NIGHT_MODE = "ui_night_mode",
    TOGGLE_POWER_SAVING_MODE = "low_power",
    SHOW_POINTER_LOCATION = "pointer_location"
}